refactor: move workspace tests into submodule
